requirements.txt
# Web scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0

# AI/ML
anthropic==0.34.2

# Database (removed sqlite3-to-mysql as it's problematic)
# SQLite3 comes built-in with Python

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2  # Updated for Python 3.12 compatibility

# Web framework (for dashboard)
flask==3.0.0
flask-cors==4.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
